no netwprk and no data found exception handle

+++ app/src/main/java/com/example/explorer_kotlin/repository/ResultRepository.kt
@@ -1,6 +1,8 @@
 package com.example.explorer_kotlin.repository
 
+import android.content.Context
 import android.util.Log
+import android.util.NoSuchPropertyException
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.Transformations
 import com.example.explorer_kotlin.database.ResultDatabase
@@ -10,8 +12,10 @@ import com.example.explorer_kotlin.model.NetworkResultContainer
 import com.example.explorer_kotlin.model.SpaceResponse
 import com.example.explorer_kotlin.model.asDatabaseModel
 import com.example.explorer_kotlin.network.NasaApi
+import com.example.explorer_kotlin.network.Utils
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
+import java.lang.Exception
 
 class ResultRepository (private val database: ResultDatabase){
 
@@ -24,10 +28,12 @@ class ResultRepository (private val database: ResultDatabase){
         it.asDomainModel()
     }
 
-    suspend fun refreshResults(query: String?, startYear: String?, endYear:String?)
+    suspend fun refreshResults(query: String?, startYear: String?, endYear:String?, context: Context)
     {
         Log.d("ResultRepository","inside refreshVideos()")
         Log.d("ResultRepository","query : ${query}, startYear: $startYear, endYear: $endYear")
+        if(!Utils.isNetworkConnected(context))
+            throw Exception("no network")
         withContext(Dispatchers.IO){
 
             val response = NasaApi.apiResponse.getSearchResults(query,
@@ -35,6 +41,11 @@ class ResultRepository (private val database: ResultDatabase){
                     endYear)
             Log.d("ResultRepository","response: $response")
             val networkContainer = NetworkResultContainer(response.collection.items)
+            if(networkContainer.itemList.isEmpty())
+            {
+                throw NoSuchPropertyException("no data found")
+            }
+
             database.itemDao.deleteAll()
             database.itemDao.insertAll(networkContainer.asDatabaseModel())
         }
